Fixed the scripts to perform calculations for Navmesh in EnemeyTankController, setting waypoints in Controller Constructor
